requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
click==8.1.3
    # via flask
flask==2.2.3
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.0
    # via -r requirements.in
slack-bolt==1.16.2
    # via -r requirements.in
slack-sdk==3.20.0
    # via slack-bolt
werkzeug==2.2.3
    # via flask
